+++ prometheus_client_shelf/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.6.0
+
+- Introduce a `collectCallback` into every metric type that allows to update the metric before the sample values are
+  collected. This is useful to perform more complex metric calculation only when the metrics are scraped.
+- **Breaking Change**: Make `collect()` method on metrics async. Related code is now async, too.
+- Polish documentation towards a `1.0.0` release.
+
 ## 0.5.1
 
 - Migrate from `pedantic` to `lints`